+++ libOTe/TwoChooseOne/KosDotExtReceiver.cpp
@@ -20,7 +20,7 @@ namespace osuCrypto
         //mCode.random(prng, baseOTs.size(), 128);
 
         mGens.resize(baseOTs.size());
-        for (u64 i = 0; i < baseOTs.size(); i++)
+        for (u64 i = 0; i <u64(baseOTs.size()); i++)
         {
             mGens[i][0].SetSeed(baseOTs[i][0]);
             mGens[i][1].SetSeed(baseOTs[i][1]);
